add unit tests for the new actor credentials removal service
