+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Tests/Handlers/RemoveActorCredentialsHandlerTests.cs
@@ -57,71 +57,4 @@ public sealed class RemoveActorCredentialsHandlerTests
         // Act + Assert
         await Assert.ThrowsAsync<NotFoundValidationException>(() => target.Handle(command, CancellationToken.None));
     }
-
-    [Fact]
-    public async Task Handle_ActorHasNoCertificateCredentials_ReturnsOk()
-    {
-        // Arrange
-        var actorRepositoryMock = new Mock<IActorRepository>();
-        var domainEventRepositoryMock = new Mock<IDomainEventRepository>();
-        var actorCredentialsRemovalService = new Mock<IActorCredentialsRemovalService>();
-        var target = new RemoveActorCredentialsHandler(
-            actorRepositoryMock.Object,
-            UnitOfWorkProviderMock.Create(),
-            domainEventRepositoryMock.Object,
-            actorCredentialsRemovalService.Object);
-
-        var actorId = Guid.NewGuid();
-        var actor = TestPreparationModels.MockedActor(actorId);
-
-        actorRepositoryMock
-            .Setup(actorRepository => actorRepository.GetAsync(actor.Id))
-            .ReturnsAsync(actor);
-
-        var command = new RemoveActorCredentialsCommand(actorId);
-
-        // Act
-        var exception = await Record.ExceptionAsync(() => target.Handle(command, CancellationToken.None));
-
-        // Assert
-        Assert.Null(exception);
-    }
-
-    [Fact]
-    public async Task Handle_ActorHasValidCertificateCredentials_ReturnsOk()
-    {
-        // Arrange
-        var actorRepositoryMock = new Mock<IActorRepository>();
-        var domainEventRepositoryMock = new Mock<IDomainEventRepository>();
-        var actorCredentialsRemovalService = new Mock<IActorCredentialsRemovalService>();
-        var target = new RemoveActorCredentialsHandler(
-            actorRepositoryMock.Object,
-            UnitOfWorkProviderMock.Create(),
-            domainEventRepositoryMock.Object,
-            actorCredentialsRemovalService.Object);
-
-        var actorId = Guid.NewGuid();
-        var actor = TestPreparationModels.MockedActor(actorId);
-        actor.Credentials = new ActorCertificateCredentials(
-            "mocked",
-            "mocked",
-            DateTime.UtcNow.AddYears(1).ToInstant());
-
-        actorRepositoryMock
-            .Setup(actorRepository => actorRepository.GetAsync(actor.Id))
-            .ReturnsAsync(actor);
-
-        actorRepositoryMock
-            .Setup(actorRepository => actorRepository.AddOrUpdateAsync(actor))
-            .ReturnsAsync(new Result<ActorId, ActorError>(actor.Id));
-
-        var command = new RemoveActorCredentialsCommand(actorId);
-
-        // Act
-        var exception = await Record.ExceptionAsync(() => target.Handle(command, CancellationToken.None));
-
-        // Assert
-        Assert.Null(exception);
-        Assert.Null(actor.Credentials);
-    }
 }
